requirements.txt
annotated-types==0.7.0
anyio==4.7.0
boto3==1.35.86
botocore==1.35.86
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
jmespath==1.0.1
protobuf==5.29.2
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
ruff==0.8.4
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
temporalio==1.8.0
types-protobuf==5.29.1.20241207
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
